Move `polyclang/polytracker.h` -> `polytracker/polytracker.h`
