+++ polytracker/src/CMakeLists.txt
@@ -1,4 +1,3 @@
-#add_subdirectory(polyclang)
 add_subdirectory(dfsan_rt)
 add_subdirectory(dfsan_pass)
 add_subdirectory(dfsan_sources)
